// src/app/modules/configuration/general/models/update-class-period-type-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * ClassPeriodTypesView
 */
export interface UpdateClassPeriodTypeView { 
    /**
     * Gets or sets Name
     */
    name: string;
    /**
     * Gets or sets IsBreak
     */
    isBreak: number;
    /**
     * Gets or sets HasMultipleAssignments
     */
    hasMultipleAssignments: number;
    /**
     * Gets or sets HasAssignments
     */
    hasAssignments: number;

    id: string;
}